Initial runnable port
- Game compiles and runs on MacOS test-machine.
- Bitmaps load successfully with magenta transparent
  color key. Antialiasing has also been fixed.
- Sprite bitmask shadows work correctly.
- SDL_Input on main menu has been fixed.

+++ src/CVehicle.cpp
@@ -1173,7 +1173,7 @@ void CVehicle::Update()
 	center_x = x + (width >> 1);
 	center_y = y + (height >> 1);
 
-	SDL_Rect rect = {0, 0, width, height};
+	SDL_Rect rect = {0, 0, (int)width, (int)height};
 
 	/*
 	 * Update bonuses states
